+++ catchup/fetcher_test.go
@@ -848,59 +848,3 @@ func makeTestUnicastPeer(gn network.GossipNode, version string, t *testing.T) ne
 	return &wsp
 }
 
-// A quick GetBlock over websockets test hitting a mocked websocket server (no actual connection)
-func TestGetBlockWS(t *testing.T) {
-	// test the WS fetcher:
-	// 1. fetcher sends UniEnsBlockReqTag to http peer
-	// 2. peer send message to gossip node
-	// 3. gossip node send message to ledger service
-	// 4. ledger service responds with UniCatchupResTag sending it back to the http peer
-	// 5. the http peer send it to the network
-	// 6. the network send it back to the fetcher
-
-	// start server
-	ledger, next, b, err := buildTestLedger(t)
-	if err != nil {
-		t.Fatal(err)
-		return
-	}
-
-	cfg := config.GetDefaultLocal()
-
-	versions := []string{"2.1"}
-	for _, version := range versions { // range network.SupportedProtocolVersions {
-
-		net := &httpTestPeerSource{}
-		blockServiceConfig := config.GetDefaultLocal()
-		blockServiceConfig.CatchupParallelBlocks = 5
-		blockServiceConfig.EnableBlockService = true
-		ls := rpcs.MakeBlockService(blockServiceConfig, ledger, net, "test genesisID")
-
-		ls.Start()
-
-		up := makeTestUnicastPeer(net, version, t)
-		net.peers = append(net.peers, up)
-
-		_, ok := net.GetPeers(network.PeersConnectedIn)[0].(network.UnicastPeer)
-		require.True(t, ok)
-		factory := MakeNetworkFetcherFactory(net, numberOfPeers, &cfg)
-		factory.log = logging.TestingLog(t)
-		fetcher := factory.NewOverGossip()
-		// we have one peer, the Ws block server
-		require.Equal(t, fetcher.NumPeers(), 1)
-
-		var block *bookkeeping.Block
-		var cert *agreement.Certificate
-		var client FetcherClient
-
-		block, cert, client, err = fetcher.FetchBlock(context.Background(), next)
-		require.NotNil(t, client)
-		require.NoError(t, err)
-		require.Equal(t, &b, block)
-		if err == nil {
-			require.NotEqual(t, nil, block)
-			require.NotEqual(t, nil, cert)
-		}
-		fetcher.Close()
-	}
-}
